fix: change age field type (Int -> Double) in Animal.kt to fix bug

+++ app/src/main/java/org/rsschool/rsandroidtask4/data/Animal.kt
@@ -8,6 +8,6 @@ data class Animal(
     @PrimaryKey(autoGenerate = true)
     val id: Int = 0,
     val name: String,
-    val age: Int,
+    val age: Double,
     val breed: String
 )
